+++ lib/dial/middleware/panel.rb
@@ -80,6 +80,7 @@ module Dial
       def style
         <<~CSS
           #dial {
+            all: initial;
             max-height: 50%;
             max-width: 50%;
             z-index: 9999;
@@ -151,19 +152,19 @@ module Dial
 
       def script
         <<~JS
-          var dialPreview = document.getElementById("dial-preview");
-          var dialDetails = document.getElementById("dial-details");
+          const dialPreview = document.getElementById("dial-preview");
+          const dialDetails = document.getElementById("dial-details");
 
           dialPreview.addEventListener("click", () => {
-            var collapsed = ["", "none"].includes(dialDetails.style.display);
-            dialDetails.style.display = collapsed ? "block" : "none";
+            const isCollapsed = ["", "none"].includes(dialDetails.style.display);
+            dialDetails.style.display = isCollapsed ? "block" : "none";
           });
 
           document.addEventListener("click", (event) => {
             if (!dialPreview.contains(event.target) && !dialDetails.contains(event.target)) {
               dialDetails.style.display = "none";
 
-              var detailsElements = dialDetails.querySelectorAll("details");
+              const detailsElements = dialDetails.querySelectorAll("details");
               detailsElements.forEach(detail => {
                 detail.removeAttribute("open");
               });