+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Prevent style inheritance in panel
+
 ## [0.2.6] - 2025-04-05
 
 - Decrease default vernier allocation interval from 20k to 2k